+++ src-golang/model/camera/camera.go
@@ -3,12 +3,12 @@ package camera
 import "src-golang/model/optics"
 
 type Camera interface {
-	GenerateRay(res *optics.Ray, index ...int64) *optics.Ray
+	GenerateRay(res *optics.Ray, index ...int) *optics.Ray
 }
 
 type CameraBase struct {
 }
 
-func (c *CameraBase) GenerateRay(ray *optics.Ray, index ...int64) *optics.Ray {
+func (c *CameraBase) GenerateRay(ray *optics.Ray, index ...int) *optics.Ray {
 	return &optics.Ray{}
 }
