+++ src-golang/ray_tracing/trace_pixel.go
@@ -9,7 +9,7 @@ import (
 )
 
 // TracePixel 追踪单个像素
-func (h *Handler) TracePixel(camera camera.Camera, objTree *object.ObjectTree, samples int64, index ...int64) *mat.VecDense {
+func (h *Handler) TracePixel(camera camera.Camera, objTree *object.ObjectTree, samples int64, index ...int) *mat.VecDense {
 	color := mat.NewVecDense(3, nil)
 	ray := h.RayPool.Get().(*optics.Ray) // new ray
 	defer h.RayPool.Put(ray)
